Actuator class
Added a new class called actuator. This class uses the LS7366R class, the AccelStepper library, and Arduino SPI library (LS7366R uses it). The purpose of the actuator class is to represent and control a stepper motor that has a quadrature optical rotary encoder on it, effectively making it an actuator. Everything is encapsulated and passed by reference.